# backend/docs/swagger.yaml
basePath: /
definitions:
  dto.AiSuggestionResponse:
    properties:
      confidence:
        type: number
      createdAt:
        type: string
      flagId:
        type: string
      id:
        type: string
      needsReview:
        type: boolean
      payload:
        type: object
      subflagId:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  dto.ConfirmInboxItemRequest:
    properties:
      flagId:
        type: string
      payload:
        type: object
      subflagId:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  dto.ConfirmInboxItemResponse:
    properties:
      event:
        $ref: '#/definitions/dto.EventResponse'
      reminder:
        $ref: '#/definitions/dto.ReminderResponse'
      shoppingItems:
        items:
          $ref: '#/definitions/dto.ShoppingItemResponse'
        type: array
      shoppingList:
        $ref: '#/definitions/dto.ShoppingListResponse'
      task:
        $ref: '#/definitions/dto.TaskResponse'
      type:
        type: string
    type: object
...
  dto.CreateInboxItemRequest:
    properties:
      rawMediaUrl:
        type: string
      rawText:
        type: string
      source:
        type: string
    type: object
  dto.CreateReminderRequest:
    properties:
      remindAt:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
  dto.EventResponse:
    properties:
      allDay:
        type: boolean
      createdAt:
        type: string
      endAt:
        type: string
      id:
        type: string
      location:
        type: string
      sourceInboxItemId:
        type: string
      startAt:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.InboxItemResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastError:
        type: string
      rawMediaUrl:
        type: string
      rawText:
        type: string
      source:
        type: string
      status:
        type: string
      suggestion:
        $ref: '#/definitions/dto.AiSuggestionResponse'
      updatedAt:
        type: string
    type: object
...
  dto.ListInboxItemsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.InboxItemResponse'
        type: array
      nextCursor:
        type: string
    type: object
  dto.ListRemindersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ReminderResponse'
        type: array
      nextCursor:
        type: string
    type: object
...
  dto.ReminderResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      remindAt:
        type: string
      sourceInboxItemId:
        type: string
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ShoppingItemResponse:
    properties:
      checked:
        type: boolean
      createdAt:
        type: string
      id:
        type: string
      listId:
        type: string
      quantity:
        type: string
      sortOrder:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ShoppingListResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      sourceInboxItemId:
        type: string
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.TaskResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      dueAt:
        type: string
      id:
        type: string
      sourceInboxItemId:
        type: string
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.UpdateReminderRequest:
    properties:
      remindAt:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: API do MVP Inbota.
  title: Inbota API
  version: "0.1"
paths:
...
  /v1/inbox-items:
    get:
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      - description: Source
        in: query
        name: source
        type: string
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListInboxItemsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar inbox items
      tags:
      - Inbox
    post:
      consumes:
      - application/json
      parameters:
      - description: Inbox payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInboxItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.InboxItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar inbox item
      tags:
      - Inbox
  /v1/inbox-items/{id}:
    get:
      parameters:
      - description: Inbox item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InboxItemResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter inbox item
      tags:
      - Inbox
  /v1/inbox-items/{id}/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: Inbox item ID
        in: path
        name: id
        required: true
        type: string
      - description: Confirm payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConfirmInboxItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ConfirmInboxItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirmar inbox item
      tags:
      - Inbox
  /v1/inbox-items/{id}/dismiss:
    post:
      parameters:
      - description: Inbox item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InboxItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Descartar inbox item
      tags:
      - Inbox
  /v1/inbox-items/{id}/reprocess:
    post:
      parameters:
      - description: Inbox item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InboxItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reprocessar inbox item
      tags:
      - Inbox
  /v1/reminders:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListRemindersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar lembretes
      tags:
      - Reminders
    post:
      consumes:
      - application/json
      parameters:
      - description: Reminder payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReminderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar lembrete
      tags:
      - Reminders
  /v1/reminders/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Reminder ID
        in: path
        name: id
        required: true
        type: string
      - description: Reminder payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReminderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReminderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar lembrete
      tags:
      - Reminders
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"